+++ scripts/fiddle-update.sh
@@ -11,6 +11,7 @@
 #  Revision History::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::|
 # ---------------------------------------------------------------------------------------------------|
 # 05/26/2018 - Baseline ~ See CHANGELOG @ 230_update_and_shrinkwrap
+# 08/01/2018 - See CHANGELOG @ 006_fiddle_react
 # ---------------------------------------------------------------------------------------------------|
 source bin/_utils.sh;
 source bin/_types.sh;
@@ -57,6 +58,11 @@ function updateFiddle() {
         'node')
             source bin/node/_install.sh;
             source bin/node/_update.sh;
+            update ${_fiddleDir};
+            ;;
+        'react')
+            source bin/react/_install.sh;
+            source bin/react/_update.sh;
             update ${_fiddleDir};
             ;;
         *)  exit 86